requirements/common.txt
Django==1.10.4
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
django-filter==1.0.1
